Move JSON parsing logic to ClusterConfig
